extracted HoneyBadgerComponent as base class
